refactor asset upload wizard components
